+++ record/CHANGELOG.md
@@ -1,4 +1,9 @@
-## 4.2.0-beta.1
+## 4.2.0-dev.2
+* feat: Add recorder state streams.
+* feat: Add recorder amplitude stream.
+* core: Update dependencies.
+
+## 4.2.0-dev.1
 * feat: Add recorder state streams.
 * feat: Add recorder amplitude stream.
 * core: Update dependencies.